# openapi.yaml
openapi: 3.0.3
info:
  title: GastronomIQ API
  description: |
    Recipe Development Platform API for culinary professionals.
    
    ## Authentication
    All endpoints (except `/auth/register` and `/auth/login`) require JWT bearer token in Authorization header.
    
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Multi-Tenancy
    Organization context is extracted from JWT claims. All data is automatically scoped to the user's organization.
    
    ## Error Handling
    All errors follow a standardized format with error codes and detailed messages.
    
  version: "1.0.0"
  contact:
    name: Support
    email: support@gastronomiq.com
  license:
    name: MIT
    url: https://github.com/CulinariaProfessionali/GastronomIQ/blob/main/LICENSE
servers:
  - url: https://api.gastronomiq.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        is_active:
          type: boolean
        email_verified:
          type: boolean
        two_factor_enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
          example: ["editor", "viewer"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Ingredient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        unit:
          type: string
          example: g
        is_allergen:
          type: boolean
        allergens:
          type: array
          items:
            type: string
          example: ["nuts", "dairy"]
        cost_per_unit:
          type: number
          format: decimal
          example: 2.50
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateIngredientRequest:
      type: object
      required:
        - name
        - unit
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        category:
          type: string
        unit:
          type: string
          enum: [g, kg, ml, l, cup, tbsp, tsp, piece, oz, lb]
        is_allergen:
          type: boolean
        allergens:
          type: array
          items:
            type: string
        cost_per_unit:
          type: number
          format: decimal
    IngredientResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Ingredient'
    IngredientListResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Recipe Schemas
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
        cuisine:
          type: string
          nullable: true
        yield_amount:
          type: number
          format: decimal
          example: 4
        yield_unit:
          type: string
          example: servings
        prep_time_minutes:
          type: integer
          nullable: true
        cook_time_minutes:
          type: integer
          nullable: true
        rest_time_minutes:
          type: integer
          nullable: true
        total_time_minutes:
          type: integer
        status:
          type: string
          enum: [draft, review, approved, published, archived]
        dietary_restrictions:
          type: array
          items:
            type: string
          example: [vegan, gluten-free]
        equipment:
          type: array
          items:
            type: string
          example: [oven, mixer]
        version_number:
          type: integer
        is_published:
          type: boolean
        published_at:
          type: string
          format: date-time
          nullable: true
        view_count:
          type: integer
        like_count:
          type: integer
        created_by:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecipeDetail:
      allOf:
        - $ref: '#/components/schemas/Recipe'
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: '#/components/schemas/RecipeStep'
            ingredients:
              type: array
              items:
                $ref: '#/components/schemas/RecipeIngredient'
            nutrition:
              $ref: '#/components/schemas/NutritionInfo'
            cost_info:
              $ref: '#/components/schemas/RecipeCost'
    CreateRecipeRequest:
      type: object
      required:
        - name
        - difficulty
        - yield_amount
        - yield_unit
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
        cuisine:
          type: string
        yield_amount:
          type: number
          format: decimal
        yield_unit:
          type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        rest_time_minutes:
          type: integer
        dietary_restrictions:
          type: array
          items:
            type: string
        equipment:
          type: array
          items:
            type: string
    UpdateRecipeRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
        cuisine:
          type: string
        yield_amount:
          type: number
        yield_unit:
          type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        rest_time_minutes:
          type: integer
        dietary_restrictions:
          type: array
          items:
            type: string
        equipment:
          type: array
          items:
            type: string
    RecipeStep:
      type: object
      properties:
        id:
          type: string
          format: uuid
        step_number:
          type: integer
        title:
          type: string
          nullable: true
        instructions:
          type: string
        duration_minutes:
          type: integer
          nullable: true
        temperature_celsius:
          type: number
          nullable: true
        equipment_needed:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
...
    RecipeIngredient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ingredient:
          $ref: '#/components/schemas/Ingredient'
        quantity:
          type: number
          format: decimal
        unit:
          type: string
        notes:
          type: string
          nullable: true
        is_optional:
          type: boolean
        substitutes:
          type: array
          items:
            type: object
    AddRecipeIngredientRequest:
      type: object
      required:
        - ingredient_id
        - quantity
        - unit
      properties:
        ingredient_id:
          type: string
          format: uuid
        quantity:
          type: number
          format: decimal
        unit:
          type: string
        notes:
          type: string
        is_optional:
          type: boolean
    NutritionInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        per_serving:
          type: boolean
        calories:
          type: number
          format: decimal
        protein_g:
          type: number
          format: decimal
        carbohydrates_g:
          type: number
          format: decimal
        fat_g:
          type: number
          format: decimal
        saturated_fat_g:
          type: number
          format: decimal
        fiber_g:
          type: number
          format: decimal
        sugar_g:
          type: number
          format: decimal
        sodium_mg:
          type: number
          format: decimal
        calculated_at:
          type: string
          format: date-time
    RecipeCost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        total_cost:
          type: number
          format: decimal
        cost_per_serving:
          type: number
          format: decimal
        currency:
          type: string
          example: USD
        calculated_at:
          type: string
          format: date-time
    RecipeResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/RecipeDetail'
    RecipeListResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Inventory Schemas
...
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalCount:
          type: integer
          example: 145
        totalPages:
          type: integer
          example: 8
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        errorCode:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /ingredients:
    get:
      tags:
        - Ingredients
      summary: List ingredients
      operationId: listIngredients
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name
        - name: category
          in: query
          schema:
            type: string
        - name: is_allergen
          in: query
          schema:
            type: boolean
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: Ingredients list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientListResponse'

    post:
      tags:
        - Ingredients
      summary: Create ingredient
      operationId: createIngredient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIngredientRequest'
      responses:
        201:
          description: Ingredient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ingredients/{ingredientId}:
    get:
      tags:
        - Ingredients
      summary: Get ingredient details
      operationId: getIngredient
      security:
        - BearerAuth: []
      parameters:
        - name: ingredientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ingredient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        404:
          description: Ingredient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Ingredients
      summary: Update ingredient
      operationId: updateIngredient
      security:
        - BearerAuth: []
      parameters:
        - name: ingredientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIngredientRequest'
      responses:
        200:
          description: Ingredient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        404:
          description: Ingredient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Ingredients
      summary: Delete ingredient
      operationId: deleteIngredient
      security:
        - BearerAuth: []
      parameters:
        - name: ingredientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Ingredient deleted
        404:
          description: Ingredient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Recipe Endpoints
  /recipes:
    get:
      tags:
        - Recipes
      summary: List recipes
      operationId: listRecipes
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, review, approved, published, archived]
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard, expert]
        - name: cuisine
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: Recipes list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeListResponse'

    post:
      tags:
        - Recipes
      summary: Create recipe
      operationId: createRecipe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        201:
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes/{recipeId}:
    get:
      tags:
        - Recipes
      summary: Get recipe details
      operationId: getRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        404:
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Recipes
      summary: Update recipe
      operationId: updateRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeRequest'
      responses:
        200:
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        404:
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Recipes
      summary: Delete recipe
      operationId: deleteRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Recipe deleted
        404:
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /recipes/{recipeId}/ingredients:
    post:
      tags:
        - Recipes
      summary: Add ingredient to recipe
      operationId: addRecipeIngredient
      security:
        - BearerAuth: []
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRecipeIngredientRequest'
      responses:
        201:
          description: Ingredient added
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/RecipeIngredient'
...
security:
  - BearerAuth: []